Move Model class into models.py
